Order Items handling and Cart checkout functionality

+++ order-service/src/main/java/com/github/dimitryivaniuta/orderservice/model/OrderStatus.java
@@ -6,30 +6,30 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 
 public enum OrderStatus {
-    PENDING(0),
-    AWAITING_PAYMENT(1),
-    RESERVED(2),
-    PAID(3),
-    REJECTED(4),
-    CANCELLED(5);
+    PENDING((short) 0),
+    AWAITING_PAYMENT((short) 1),
+    RESERVED((short) 2),
+    PAID((short) 3),
+    REJECTED((short) 4),
+    CANCELLED((short) 5);
 
-    private final int code;
+    private final short code;
 
-    private static final Map<Integer, OrderStatus> LOOKUP =
+    private static final Map<Short, OrderStatus> LOOKUP =
             Arrays.stream(values())
                     .collect(Collectors.toUnmodifiableMap(OrderStatus::code, Function.identity()));
 
-    OrderStatus(int  code) {
+    OrderStatus(short  code) {
         this.code = code;
     }
 
     /** Numeric code stored in DB (SMALLINT). */
-    public int  code() {
+    public short  code() {
         return code;
     }
 
     /** Resolve from a SMALLINT value. Throws if unknown. */
-    public static OrderStatus fromCode(int code) {
+    public static OrderStatus fromCode(short code) {
         OrderStatus status = LOOKUP.get(code);
         if (status == null) {
             throw new IllegalArgumentException("Unknown OrderStatus code: " + code);
