+++ order-service/src/main/java/com/github/dimitryivaniuta/orderservice/model/converter/OrderStatusRead.java
@@ -5,6 +5,6 @@ import org.springframework.core.convert.converter.Converter;
 import org.springframework.data.convert.ReadingConverter;
 
 @ReadingConverter
-public class OrderStatusRead implements Converter<Integer, OrderStatus> {
-    public OrderStatus convert(Integer source){ return OrderStatus.fromCode(source); }
+public class OrderStatusRead implements Converter<Short , OrderStatus> {
+    public OrderStatus convert(Short  source){ return OrderStatus.fromCode(source); }
 }
